+++ kaithem/src/docs/changes.md
@@ -10,7 +10,7 @@ Change Log
 * [sparkles] Can use numbered midi notes like midi.note:1.60 in chandler
 * [sparkles] KAITHEM_AUDIO_API=jack sets the mixer to use jack instead of direct pipewire.
 * [bug] Work around gstreamer struct version regression
-
+* [coffin] Remove sdmon support
 
 ### 0.96.0